% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGSEA.R
\name{plotGSEA}
\alias{plotGSEA}
\title{Plot for gene enrichment analysis of GSEA method}
\usage{
plotGSEA(
  gsea_list,
  plot_type = c("volcano", "classic", "fgsea", "ridge", "bar"),
  stats_metric = c("p.adjust", "pvalue", "qvalue"),
  show_pathway = NULL,
  show_gene = NULL,
  colour = NULL,
  wrap_length = NULL,
  label_by = c("id", "description"),
  ...
)
}
\arguments{
\item{gsea_list}{GSEA result from `genGSEA` function}

\item{plot_type}{GSEA plot type, one of 'volcano', 'classic', 'fgsea', 'ridge' or 'bar'.}

\item{stats_metric}{Statistic metric from one of "pvalue", "p.adjust", "qvalue".}

\item{show_pathway}{Select plotting pathways by number (will choose top N pathways)
or pathway name (choose from ID column).}

\item{show_gene}{Select genes to show. Default is "all". Used in "classic" plot.}

\item{colour}{Colour vector. Deafault is NULL. Used in volcano, ridge and bar plot.}

\item{wrap_length}{Numeric, wrap text if longer than this length. Default is NULL.}

\item{label_by}{Select which column as the label. If user wants to modify labels in plot,
please modify the "Description" column and set the argument as "description". Default is by 'id'.}

\item{...}{other arguments transfer to `plot_theme` function}
}
\value{
A ggplot object
}
\description{
Gene Set Enrichment Analysis (GSEA) is a computational method that determines whether an a priori defined set of
genes shows statistically significant, concordant differences between two biological states (e.g. phenotypes).
}
